backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser==2.2.2
gunicorn==20.1.0
python-dotenv==1.0.0
Pillow==10.4.0
django-filter==23.2
drf-extra-fields==3.7.0
